refactor: select component folder move
