+++ apps/web/src/components/FeedBack/FeedBack.tsx
@@ -2,8 +2,9 @@
 
 import styled from 'styled-components';
 
+import Select from '@/components/Select';
+
 import { ISSUE_LABEL } from './FeedBack.constants';
-import Select from './Select';
 
 function FeedBack() {
   return (
